Factor out WiiMote MegaViewport class
