+++ lg_pointer/CMakeLists.txt
@@ -5,6 +5,8 @@ find_package(catkin REQUIRED COMPONENTS
   rospy
 )
 
+catkin_python_setup()
+
 catkin_package(
   CATKIN_DEPENDS
   rospy
