+++ lg_pointer/src/lg_pointer/__init__.py
@@ -1,0 +1,1 @@
+from megaviewport import MegaViewport
